requirements.txt
django
djangorestframework
pymilvus
requests
environ
psycopg2-binary
django-cors-headers